+++ Capivara/DesignSystem/Components/RoundedCorner.swift
@@ -1,0 +1,19 @@
+//
+//  RoundedCorner.swift
+//  Capivara
+//
+//  Created by Vitor Cheung on 06/12/21.
+//
+
+import SwiftUI
+
+struct RoundedCorner: Shape{
+    
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+    
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
