+++ Capivara/Screen/MyEvents/Views/EventCardView.swift
@@ -63,23 +63,6 @@ struct EventCardView: View {
     }
 }
 
-struct RoundedCorner: Shape{
-    
-    var radius: CGFloat = .infinity
-    var corners: UIRectCorner = .allCorners
-    
-    func path(in rect: CGRect) -> Path {
-        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
-        return Path(path.cgPath)
-    }
-}
-
-extension View{
-    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View{
-        clipShape(RoundedCorner(radius: radius, corners: corners))
-    }
-}
-
 struct EventCardView_Previews: PreviewProvider {
     static var previews: some View {
         TabView {
